+++ sc-kpm/python/sc_main.py
@@ -1,3 +1,0 @@
-import sys, os
-
-print("Hello World")