+++ sc-memory/cpp/python/sc_python_module.cpp
@@ -4,6 +4,12 @@
 #include "../sc_stream.hpp"
 #include "../sc_link.hpp"
 
+extern "C"
+{
+#include "sc-memory/sc_memory_headers.h"
+}
+
+
 namespace
 {
 namespace bp = boost::python;
@@ -248,6 +254,11 @@ uint8_t PyLinkContent::Type::Int = 1;
 uint8_t PyLinkContent::Type::Float = 2;
 
 // ----------------------------
+
+ScMemoryContext * _context_CreateInstance(std::string const & name)
+{
+  return new ScMemoryContext(sc_access_lvl_make_min, name.c_str());
+}
 
 bp::list _context_FindLinksByContent(ScMemoryContext & self, bp::object const & content)
 {
@@ -695,14 +706,15 @@ bp::object _context_helperBuildTemplate(ScMemoryContext & self, ScAddr const & t
   return bp::object();
 }
 
-ScMemoryContext * _context_new(std::string const & name)
-{
-  return new ScMemoryContext(sc_access_lvl_make_min, name);
-}
-
 bp::object ScAddrFromHash(ScAddr::HashType const value)
 {
   return bp::object(ScAddr(value));
+}
+
+std::string GetConfigValue(std::string const & group, std::string const & key)
+{
+  char const * value = sc_config_get_value_string(group.c_str(), key.c_str());
+  return value ? std::string(value) : "";
 }
 
 } // namespace impl
@@ -711,10 +723,12 @@ BOOST_PYTHON_MODULE(sc)
 {
   bp::register_exception_translator<utils::ScException>(&translateException);
 
-  def("createScMemoryContext", bp::make_function(&impl::_context_new, bp::return_value_policy<bp::manage_new_object>()));
   def("ScAddrFromHash", bp::make_function(&impl::ScAddrFromHash));
+  def("getScConfigValue", bp::make_function(&impl::GetConfigValue));
 
   bp::class_<ScMemoryContext, boost::noncopyable>("ScMemoryContext", bp::no_init)
+    .def("Create", &impl::_context_CreateInstance, bp::return_value_policy<bp::manage_new_object>())
+    .staticmethod("Create")
     .def("CreateNode", &ScMemoryContext::CreateNode, bp::return_value_policy<bp::return_by_value>())
     .def("CreateEdge", &ScMemoryContext::CreateEdge)
     .def("CreateLink", &ScMemoryContext::CreateLink)
@@ -811,7 +825,7 @@ BOOST_PYTHON_MODULE(sc)
     .def("IsLink", &ScType::IsLink)
     .def("IsEdge", &ScType::IsEdge)
     .def("IsNode", &ScType::IsNode)
-    .def("IsValid", &ScType::IsValid)
+    .def("IsUnknown", &ScType::IsUnknown)
     .def("IsConst", &ScType::IsConst)
     .def("IsVar", &ScType::IsVar)
     .def("ToInt", &ScType::operator*)