+++ sc-memory/cpp/python/sc_python_threads.hpp
@@ -32,4 +32,29 @@ private:
   PyGILState_STATE m_subGILState;
 };
 
+class WithoutGIL
+{
+public:
+  WithoutGIL() { m_state = PyEval_SaveThread(); }
+  ~WithoutGIL() { PyEval_RestoreThread(m_state); }
+
+  WithoutGIL(WithoutGIL const &) = delete;
+  WithoutGIL & operator=(WithoutGIL const &) = delete;
+private:
+  PyThreadState * m_state;
+};
+
+class WithGIL
+{
+public:
+  WithGIL() { m_state = PyGILState_Ensure(); }
+  ~WithGIL() { PyGILState_Release(m_state); }
+
+  WithGIL(WithoutGIL const &) = delete;
+  WithGIL & operator=(WithoutGIL const &) = delete;
+
+private:
+  PyGILState_STATE m_state;
+};
+
 } // namespace py