+++ sc-memory/cpp/scs/scs_types.hpp
@@ -17,9 +17,12 @@ class TypeResolver final
 public:
   _SC_EXTERN static ScType const & GetConnectorType(std::string const & connectorAlias);
   _SC_EXTERN static ScType const & GetKeynodeType(std::string const & keynodeAlias);
+
   _SC_EXTERN static bool IsConnectorReversed(std::string const & connectorAlias);
   _SC_EXTERN static bool IsConst(std::string const & idtf);
   _SC_EXTERN static bool IsEdgeAttrConst(std::string const & attr);
+  _SC_EXTERN static bool IsKeynodeType(std::string const & alias);
+  _SC_EXTERN static bool IsUnnamed(std::string const & alias);
 
 private:
   using MapType = std::unordered_map<std::string, ScType>;